Added RewardValue class to RewardsCovertor On branch reward-value Changes not staged for commit:   (use "git add/rm <file>..." to update what will be committed)   (use "git restore <file>..." to discard changes in working directory)         deleted:    src/main/java/RewardsConverter.java         deleted:    src/test/java/RewardValueTests.java
Untracked files:
  (use "git add <file>..." to include in what will be committed)
        src/main/java/RewardsConverter$RewardValue.class
        src/main/java/RewardsConverter.class
        src/main/java/src/
        src/test/java/src/

no changes added to commit (use "git add" and/or "git commit -a")
